add class Intro and curtain

+++ 05-SceneManager/Game.h
@@ -61,7 +61,7 @@ class CGame
 	void _ParseSection_SCENES(string line);
 
 public:
-	int nextX = 0, nextY = 0;
+	float nextX = 0.0f, nextY = 0.0f;
 	// Init DirectX, Sprite Handler
 	void Init(HWND hWnd, HINSTANCE hInstance);
